qt/requirements.txt
aio-pika==9.5.0
aiormq==6.8.1
exceptiongroup==1.2.2
grpcio==1.68.0
grpcio-tools==1.68.0
idna==3.10
multidict==6.1.0
pamqp==3.3.0
pika==1.3.2
propcache==0.2.0
protobuf==5.28.3
PyQt5==5.15.11
PyQt5-Qt5==5.15.15
PyQt5_sip==12.15.0
PyYAML==6.0.2
typing_extensions==4.12.2
yarl==1.18.0
